Change image text retrieval to feature extraction
